requirements.txt
streamlit
pandas
numpy
smolagents
python-dotenv
litellm